+++ libkbfs/journal_server.go
@@ -171,6 +171,31 @@ func (j *JournalServer) Disable(tlfID TlfID) (err error) {
 	return nil
 }
 
+func (j *JournalServer) blockServer() journalBlockServer {
+	return journalBlockServer{j, j.delegateBlockServer}
+}
+
+func (j *JournalServer) mdOps() journalMDOps {
+	return journalMDOps{j.delegateMDOps, j, j.delegateMDServer}
+}
+
+func makeJournalServer(
+	config Config, log logger.Logger,
+	dir string, bserver BlockServer,
+	mdOps MDOps, mdServer MDServer) *JournalServer {
+	jServer := JournalServer{
+		config:              config,
+		log:                 log,
+		deferLog:            log.CloneWithAddedDepth(1),
+		dir:                 dir,
+		delegateBlockServer: bserver,
+		delegateMDOps:       mdOps,
+		delegateMDServer:    mdServer,
+		tlfBundles:          make(map[TlfID]*tlfJournalBundle),
+	}
+	return &jServer
+}
+
 type journalBlockServer struct {
 	jServer *JournalServer
 	BlockServer
